requirements.txt
bcrypt==3.1.7
cffi==1.13.2
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.8.1
gunicorn==20.0.4
Pillow==6.2.1
psycopg2-binary==2.8.4
pycparser==2.19
pytz==2019.2
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4